Group conjured item tests
